Cancellation options, improved threads, success information display
Installed a cancellation to the scan option with improved thread
handling. Termination takes a maximum of 5-10 seconds and does not allow
a new scan to initiate while termination is taking place. 

If a distributed attack is successful the client is notified. This is
conducted by a intermittent poll to the server to find the result.
Catches need to be put in place for when connection is lost, this has
been an issue. 

GUI has been playing up, needs an upgrade. OK as proto but final should
be far more appealing/smooth. Paint commands seem to be editing
incorrect rectangles, needs addressing.