Version 1.5
Client now has an options screen in order to allow change from brute to
dictionary attack methods as well as uploading a line separated text
word list to be uploaded to the associated database for use by the
attack nodes.  

Handles the use of the dictionary method as well as checking for content
before issuing attack method. Essentially the system ensures that there
is content before allowing the user to issue an attack while also
allowing them to upload new word lists as and when they want through the
options screen. 

Full monitoring of server connection and also checks for DB connection
before allowing attack command.